+++ src/compile/headers.c
@@ -5,8 +5,8 @@
 #include "../ast.h"
 #include "../config.h"
 #include "../environment.h"
-#include "../modules.h"
 #include "../naming.h"
+#include "../packages.h"
 #include "../stdlib/datatypes.h"
 #include "../stdlib/paths.h"
 #include "../stdlib/text.h"
@@ -157,8 +157,8 @@ Text_t compile_statement_type_header(env_t *env, Path_t header_path, ast_t *ast)
         Path_t source_dir = Path$parent(source_path);
         Path_t build_dir = Path$resolved(Path$parent(header_path), Path$current_dir());
         switch (use->what) {
-        case USE_MODULE: {
-            OptionalPath_t installed = find_installed_module(ast);
+        case USE_PACKAGE: {
+            OptionalPath_t installed = find_installed_package(ast);
             assert(installed);
             List_t children = Path$glob(Path$child(installed, Text("/[!._0-9]*.tm")));
             Text_t includes = EMPTY_TEXT;