+++ src/packages.h
@@ -1,0 +1,10 @@
+// Logic for getting information about and installing packages
+
+#pragma once
+
+#include <stdbool.h>
+
+#include "ast.h"
+
+Text_t get_library_name(Path_t lib_dir);
+OptionalPath_t find_installed_package(ast_t *use);